requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
codecov>=1.4.0
pytest-cov
pytest==3.4.1
pytest-runner==4.0
yapf==0.20.2
click==6.7
GitPython==2.1.8
git+https://github.com/google/pasta.git@91cbb35e288095f71e5a19fb64d083c87c7193ab#egg=google_pasta
six==1.11.0
tqdm==4.19.6
